Changed buffer creation flags to use packed structs instead of u32
This is for sure better for safety but I am not sure about the
ergonomics of it. In particular initializing the bitfield
(with `BufferUsage{ .vertex_buffer = true }`) is a bit lengthy.
It would be much better to have the syntax for doing
`BufferUsage{ .vertex_buffer, .transfer_dst }` but I guess we have
to wait for future Zig's versions...