Parametrized in-flight frames
Now the types of `vulkan.zig` are parametrized by the number of
in-flight frames. For now renderer only uses 2 frames so the code
is the same as before. However this change made size of `Device`
known at compile time so it removes the need for `Device` to
store an `Allocator`.